[APG-480] Updates in line with review comments

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsaccreditedprogrammesapi/restapi/model/ReferralStatistics.kt
@@ -1,16 +1,17 @@
 package uk.gov.justice.digital.hmpps.hmppsaccreditedprogrammesapi.restapi.model
 
 import uk.gov.justice.digital.hmpps.hmppsaccreditedprogrammesapi.domain.repository.StatisticsRepository.ReferralStatisticsProjection
+import java.math.BigInteger
 
 data class ReferralStatistics(
-  val submittedReferralCount: Long,
-  val draftReferralCount: Long,
+  val submittedReferralCount: BigInteger,
+  val draftReferralCount: BigInteger,
   val averageDuration: String,
 ) {
   companion object {
     fun from(referralStatistics: ReferralStatisticsProjection): ReferralStatistics = ReferralStatistics(
-      submittedReferralCount = referralStatistics.getSubmittedReferrals().toLong(),
-      draftReferralCount = referralStatistics.getDraftReferrals().toLong(),
+      submittedReferralCount = referralStatistics.getSubmittedReferrals(),
+      draftReferralCount = referralStatistics.getDraftReferrals(),
       averageDuration = referralStatistics.getAverageDuration(),
     )
   }
